Module InferBase.ResultsDir

val assert_results_dir : string -> unit

Check that the results dir exists and sets up logging, the database, etc.

val remove_results_dir : unit -> unit

Recursively delete the results directory.

val create_results_dir : unit -> unit

Create the results dir and sets up logging, the database, etc.

val delete_capture_and_results_data : unit -> unit

delete capture and results data in the results directory

val dirs_to_clean : cache_capture:bool -> string InferStdlib.IStd.list

directories in the results directory containing capture or results data